chatbots/chat_bot/requirements.txt
flask 
openai 
python-dotenv